gamit_app/requirements.txt
asgiref==3.9.1
Django==5.2.5
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary>=2.9
whitenoise>=6.5
gunicorn>=20.1
django-import-export
reportlab==4.1.0
pypdf==4.0.1